refactor(context)!: move provider and consumer into /color
