+++ elements/rh-accordion/rh-accordion-header.ts
@@ -1,11 +1,12 @@
 import type { TemplateResult } from 'lit';
-import type { ColorPalette, ColorTheme } from '../../lib/context/types.js';
 
 import { html } from 'lit';
 import { classMap } from 'lit/directives/class-map.js';
 import { customElement, property } from 'lit/decorators.js';
-import { colorContextConsumer } from '../../lib/context/decorators.js';
+
 import { DirController } from '../../lib/DirController.js';
+
+import { colorContextConsumer, type ColorTheme } from '../../lib/context/color/consumer.js';
 
 import { BaseAccordionHeader } from '@patternfly/pfe-accordion/BaseAccordionHeader.js';
