+++ elements/rh-cta/rh-cta.ts
@@ -1,17 +1,17 @@
-import type { ColorPalette, ColorTheme } from '../../lib/context/types.js';
-
 import { LitElement, html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
+import { classMap } from 'lit/directives/class-map.js';
 
 import { Logger } from '@patternfly/pfe-core/controllers/logger.js';
 
 import { DirController } from '../../lib/DirController.js';
-import { colorContextConsumer, colorContextProvider } from '../../lib/context/decorators.js';
-import { classMap } from 'lit/directives/class-map.js';
 
-import '@patternfly/pfe-icon';
+import { colorContextConsumer, type ColorTheme } from '../../lib/context/color/consumer.js';
+import { colorContextProvider, type ColorPalette } from '../../lib/context/color/provider.js';
 
 import style from './rh-cta.css';
+
+import '@patternfly/pfe-icon';
 
 export interface CtaData {
   href?: string;
