+++ elements/rh-footer/RhFooter.ts
@@ -1,5 +1,3 @@
-import type { ColorPalette } from '../../lib/context/types.js';
-
 import { LitElement, html } from 'lit';
 import { html as staticHtml, unsafeStatic } from 'lit/static-html.js';
 import { property } from 'lit/decorators.js';
@@ -12,7 +10,7 @@ import style from './rh-footer.css';
 import { responsiveStyles } from './rh-footer-responsive.css.js';
 
 import { tabletLandscapeBreakpoint } from '../../lib/tokens.js';
-import { colorContextProvider } from '../../lib/context/decorators.js';
+import { colorContextProvider, type ColorPalette } from '../../lib/context/color/provider.js';
 // TODO: use ScreenSizeController
 import { MatchMediaController } from '../../lib/MatchMediaController.js';
