+++ elements/rh-secondary-nav/rh-secondary-nav-dropdown.ts
@@ -8,7 +8,7 @@ import { bound, observed } from '@patternfly/pfe-core/decorators.js';
 import { SlotController } from '@patternfly/pfe-core/controllers/slot-controller.js';
 import { getRandomId } from '@patternfly/pfe-core/functions/random.js';
 
-import { colorContextProvider } from '../../lib/context/decorators.js';
+import { colorContextProvider, type ColorPalette } from '../../lib/context/color/provider.js';
 
 import { RhSecondaryNavMenu } from './rh-secondary-nav-menu.js';
 
@@ -49,7 +49,7 @@ export class RhSecondaryNavDropdown extends LitElement {
   @state() expanded = false;
 
   @colorContextProvider()
-  @property({ reflect: true, attribute: 'color-palette' }) colorPalette = 'lighter';
+  @property({ reflect: true, attribute: 'color-palette' }) colorPalette: ColorPalette = 'lighter';
 
   connectedCallback(): void {
     super.connectedCallback();
